refactor - keep schemas with their uses
