+++ handlers/mparticle-api/runManual/invokeHandlerSARStatusCODE.ts
@@ -1,8 +1,6 @@
 import { invokeFunction } from '@modules/aws/lambda';
-import {
-	BatonSarEventStatusRequest,
-	InitiationReference,
-} from '../src/routers/baton/types-and-schemas';
+import { InitiationReference } from '../src/routers/baton/initiationReference';
+import { BatonSarEventStatusRequest } from '../src/routers/baton/access/handleStatus';
 
 const initiationReference: InitiationReference =
 	'8a6747ea-cef6-4b2d-8178-8bb81b13c4e9' as InitiationReference;
