+++ handlers/mparticle-api/src/routers/http/upload-event-batch.ts
@@ -1,29 +1,39 @@
 import type { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda';
 import { z } from 'zod';
-import type { EventBatch } from '../../../interfaces/event-batch';
-import { uploadAnEventBatch } from '../../apis/events';
-import { EventsAPI, MParticleClient } from '../../apis/mparticleClient';
 
+import { EventsAPI, MParticleClient } from '../../services/mparticleClient';
+import {
+	EventBatch,
+	uploadAnEventBatch,
+} from '../../apis/events/uploadAnEventBatch';
+
+/**
+ * Event Batch
+ * https://docs.mparticle.com/developers/apis/http/#example-json-request-body
+ * https://docs.mparticle.com/developers/apis/json-reference/
+ */
 export const eventBatchParser = {
 	body: z.object({
-		events: z.array(
-			z.object({
-				/** Custom data payload for the event */
-				data: z.record(z.string(), z.unknown()),
+		events: z
+			.array(
+				z.object({
+					/** Custom data payload for the event */
+					data: z.record(z.string(), z.unknown()),
 
-				/** Type identifier for the event */
-				eventType: z.string(),
-			}),
-		),
-		deviceInfo: z.record(z.string(), z.unknown()),
+					/** Type identifier for the event */
+					eventType: z.string(),
+				}),
+			)
+			.optional(),
+		deviceInfo: z.record(z.string(), z.unknown()).optional(),
 		userAttributes: z.record(z.string(), z.unknown()),
-		deletedUserAttributes: z.array(z.string()),
-		userIdentities: z.record(z.string(), z.unknown()),
-		applicationInfo: z.record(z.string(), z.unknown()),
-		schemaVersion: z.number(),
-		environment: z.enum(['production', 'development']),
-		context: z.record(z.string(), z.unknown()),
-		ip: z.string(),
+		deletedUserAttributes: z.array(z.string()).optional(),
+		userIdentities: z.record(z.string(), z.unknown()), //
+		applicationInfo: z.record(z.string(), z.unknown()).optional(),
+		schemaVersion: z.number().optional(),
+		environment: z.enum(['production', 'development']), //
+		context: z.record(z.string(), z.unknown()).optional(),
+		ip: z.string().optional(),
 	}),
 };
