+++ handlers/mparticle-api/src/services/batonS3Writer.ts
