+++ handlers/mparticle-api/src/services/mparticleClient.ts
@@ -1,6 +1,6 @@
-import type { AppConfig } from '../utils/config';
-import type { HttpResponse, Schema } from '../utils/make-http-request';
-import { RestRequestMaker } from '../utils/make-http-request';
+import type { AppConfig } from './config';
+import type { HttpResponse, Schema } from './make-http-request';
+import { RestRequestMaker } from './make-http-request';
 
 export interface DataSubjectAPI {
 	readonly clientType: 'dataSubject';
